requirements.txt
aiohttp==3.6.2
alembic==1.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dataset==1.2.0
discord==1.0.1
discord.py==1.3.1
idna==2.8
Mako==1.1.1
MarkupSafe==1.1.1
multidict==4.7.4
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
regex==2020.1.8
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
